% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{EclBinaryParser}
\alias{EclBinaryParser}
\title{Call to Python class to convert Eclipse binary files}
\usage{
EclBinaryParser(o, file)
}
\arguments{
\item{o}{object}

\item{file}{the name of a Eclipse binary file}
}
\description{
Uses a Python package under restools to make the convertion from Eclipse
binary files to dataframes
}
\examples{
\dontrun{
ecl_folder <- system.file("rawdata", package = "rEcl")
unsmry_file <- file.path(ecl_folder, "PUNQS3", "PUNQS3.UNSMRY")
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
}
}
