% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{get_actnum}
\alias{get_actnum}
\title{Get the activated cells from the reservoir model}
\usage{
get_actnum(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Get the activated cells from the reservoir model
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
get_actnum(parser)
}
}
