% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{get_dimensions}
\alias{get_dimensions}
\title{Get the dimensions of the main dataframe read from Eclipse binary file}
\usage{
get_dimensions(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Get the dimensions of the main dataframe read from Eclipse binary file
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
get_dimensions(parser)
}
}
