% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{get_seqnum_dates}
\alias{get_seqnum_dates}
\title{Get the seqnum sequential dates}
\usage{
get_seqnum_dates(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Get the seqnum sequential dates
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
get_seqnum_dates(parser)
}
}
