% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{get_vector_column}
\alias{get_vector_column}
\title{Get a column or several columns from the main dataframe}
\usage{
get_vector_column(p, cols)
}
\arguments{
\item{p}{parser object}

\item{cols}{character vectors with names of the columns}
}
\description{
This function has been vectorized to read multiple columns
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
get_vector_column(parser, "FOPR")
# vectorized function to get several vectors at once
df_vars <- get_vector_column(parser, c("YEARS", "FGOR", "FOPR", "FWCT"))
}
}
