% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{get_vector_names}
\alias{get_vector_names}
\title{Get all the names for the vectors extracted from the main dataframe}
\usage{
get_vector_names(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Get all the names for the vectors extracted from the main dataframe
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
get_vectors_shape(parser)
}
}
