% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{get_vectors_shape}
\alias{get_vectors_shape}
\title{Get the shape or dimensions of the vectors dataframe}
\usage{
get_vectors_shape(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Get the shape or dimensions of the vectors dataframe
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
get_vectors_shape(parser)
}
}
