% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{is_dual}
\alias{is_dual}
\title{Ask if the model is of dual porosity}
\usage{
is_dual(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Ask if the model is of dual porosity
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
is_dual(parser)
}
}
