% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{read_vectors}
\alias{read_vectors}
\title{Read the vectors from the main dataframe}
\usage{
read_vectors(p)
}
\arguments{
\item{p}{parser object}
}
\description{
Read the vectors from the main dataframe
}
\examples{
\dontrun{
ecl_folder <- system.file("rawdata", package = "rEcl")
unsmry_file <- file.path(ecl_folder, "PUNQS3", "PUNQS3.UNSMRY")
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
read_vectors(parser)
}
}
