% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restools.R
\name{restools_connect}
\alias{restools_connect}
\title{Connect to Python and the package restools}
\usage{
restools_connect()
}
\description{
Connect to Python and the package restools
}
\examples{
\dontrun{
py <- restools_connect()
parser <- EclBinaryParser(py, unsmry_file)
}
}
